requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0 
openai>=1.0.0
python-dotenv>=1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.0
spacy==3.6.1
numpy==1.24.3
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0